+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#23](https://github.com/LaxarJS/laxar-tooling/issues/23): page assembler: apply feature defaults even if widget instances do not have feature configuration
+
 
 ## v0.6.0-alpha.0